move Dimension menu to WorldInfo
this allows DimensionIdentifier to be just a Dimension provider, like all other *Identifiers
