+++ worldinfo.cpp
@@ -2,12 +2,16 @@
 
 #include "worldinfo.h"
 
+#include <QtWidgets/QMenu>
+#include <QDirIterator>
+
 #include "nbt/nbt.h"
 #include "zipreader.h"
 #include "json/json.h"
 
 
 WorldInfo::WorldInfo()
+  : menuActionGroup(NULL)
 {
   clear();
 }
@@ -173,4 +177,144 @@ void WorldInfo::clear()
   spawnZ      = 0;
 
   dimensions.clear();
+}
+
+
+
+//-------------------------------------------------------------------------------------------------
+// Dimension view menu
+
+void WorldInfo::clearDimensionsMenu(QMenu *menu) {
+  for (int i = 0; i < currentMenuActions.count(); i++) {
+    menu           ->removeAction(currentMenuActions[i]);
+    menuActionGroup->removeAction(currentMenuActions[i]);
+    delete currentMenuActions[i];
+  }
+  currentMenuActions.clear();
+  foundDimensionDirs.clear();
+  menu->setEnabled(false);
+  if (menuActionGroup != NULL) {
+    delete menuActionGroup;
+    menuActionGroup = NULL;
+  }
+}
+
+void WorldInfo::getDimensionsInWorld(QDir path, QMenu *menu, QObject *parent) {
+  // first get the currently selected dimension so it doesn't change
+  int currentIdx = -1;
+  for (int i = 0; i < currentMenuActions.length(); i++)
+    if (currentMenuActions[i]->isChecked())
+      currentIdx = currentMenuActions[i]->data().toInt();
+  clearDimensionsMenu(menu);
+  menuActionGroup = new QActionGroup(parent);
+
+  // add normal Dimensions
+  int index = 0;
+  while (true) {
+    const DimensionInfo & dim = DimensionIdentifier::Instance().getDimensionInfo(index++);
+    if (dim.name == "Dummy Dimension")
+      break;
+
+    if (dim.enabled) {
+      // check path for regex
+      if (dim.pathIsRegEx) {
+        QDirIterator it(path.absolutePath(), QDir::Dirs);
+        QRegExp rx(dim.path);
+        while (it.hasNext()) {
+          it.next();
+          if (rx.indexIn(it.fileName()) != -1) {
+            QString name = dim.name;
+            for (int c = 0; c < rx.captureCount(); c++)
+              name = name.arg(rx.cap(c + 1));
+            addDimensionToMenu(path, it.fileName(), name, parent);
+          }
+        }
+      } else {
+        addDimensionToMenu(path, dim.path, dim.name, parent);
+      }
+    }
+  }
+
+  // add Custom Dimensions
+  for (auto dim: dimensions) {
+    addDimensionToMenu(path, dim.path, dim.name, parent);
+  }
+
+  // re-add new build actions to menu
+  menu->addActions(currentMenuActions);
+  if (currentMenuActions.count() > 0) {
+    bool changed = true;
+    // locate our old selected item
+    for (int i = 0; i < currentMenuActions.length(); i++) {
+      if (currentMenuActions[i]->data().toInt() == currentIdx) {
+        currentMenuActions[i]->setChecked(true);
+        changed = false;
+        break;
+      }
+    }
+    if (changed) {
+      currentMenuActions.first()->setChecked(true);
+      int idx = currentMenuActions.first()->data().toInt();
+      emit dimensionChanged(DimensionIdentifier::Instance().getDimensionInfo(idx));
+    }
+    menu->setEnabled(true);
+  }
+}
+
+
+
+#define DIM_MAGIC 0x10000
+
+void WorldInfo::addDimensionToMenu(QDir path, QString dir, QString name, QObject *parent) {
+  // prevent adding non-existing directory
+  if (!path.exists(dir))
+    return;
+
+  // prevent adding unused dimension
+  if (!path.exists(dir + "/region"))
+    return;
+
+  // prevent re-adding already found directory
+  if (foundDimensionDirs.contains(dir))
+    return;
+
+  QAction *action = new QAction(parent);
+  action->setText(name);
+  // find index in definition list
+  int index = DimensionIdentifier::Instance().getDimensionIndex(name);
+  if (index > -1) {
+    // found a matching index in normal Dimensions
+    action->setData(index);
+  } else {
+    // find index in custom Dimension list
+    WorldInfo & wi(WorldInfo::Instance());
+    const QList<DimensionInfo> & custom = wi.getDimensions();
+    for (int idx = 0; idx<custom.length(); idx++) {
+      if (custom[idx].name == name) {
+        action->setData(idx+DIM_MAGIC);
+      }
+    }
+  }
+
+  action->setCheckable(true);
+  parent->connect(action, SIGNAL(triggered()),
+                  this,   SLOT(changeViewToDimension()));
+  menuActionGroup->addAction(action);
+  currentMenuActions.append(action);
+  foundDimensionDirs.append(dir);
+}
+
+
+void WorldInfo::changeViewToDimension() {
+  QAction *action = qobject_cast<QAction*>(sender());
+  if (action) {
+    int idx = action->data().toInt();
+    if (idx < DIM_MAGIC) {
+      emit dimensionChanged(DimensionIdentifier::Instance().getDimensionInfo(idx));
+    } else {
+      WorldInfo & wi(WorldInfo::Instance());
+      const QList<DimensionInfo> & custom = wi.getDimensions();
+      emit dimensionChanged(custom[idx-DIM_MAGIC]);
+    }
+  }
 }
