+++ worldinfo.h
@@ -5,12 +5,15 @@
 #include <QString>
 #include <QDir>
 #include <QList>
+#include <QMenu>
 
 #include "identifier/dimensionidentifier.h"
 
 
-class WorldInfo
+class WorldInfo : public QObject
 {
+  Q_OBJECT
+
  public:
   // singleton: access to global usable instance
   static WorldInfo &Instance();
@@ -26,6 +29,16 @@ class WorldInfo
   int                 getSpawnZ() const      { return spawnZ; };
 
   const QList<DimensionInfo> & getDimensions() const { return dimensions; };
+
+  // Dimension view menu
+  void clearDimensionsMenu(QMenu *menu);
+  void getDimensionsInWorld(QDir path, QMenu *menu, QObject *parent);
+
+ signals:
+  void dimensionChanged(const DimensionInfo &dim);    // dimension changed in menu
+
+ private slots:
+  void changeViewToDimension();                       // dimension changed in menu
 
  private:
   // singleton: prevent access to constructor and copyconstructor
@@ -44,6 +57,13 @@ class WorldInfo
   int                   spawnZ;
 
   QList<DimensionInfo>  dimensions;
+
+  // Dimension view menu
+  void addDimensionToMenu(QDir path, QString dir, QString name, QObject *parent);
+
+  QList<QAction *> currentMenuActions;
+  QActionGroup *   menuActionGroup;
+  QList<QString>   foundDimensionDirs;  // all directories where we already found a Dimension
 };
 
 #endif // WORLDINFO_H
